+++ src/core2/code_B8020.c
@@ -15,6 +15,6 @@ BKSpriteTextureBlock *func_8033EFB0(Struct84s *arg0, s32 arg1){
         return D_80384FF0;
     }
     
-    mem_ptr = *offset_ptr + (s32) &arg0[0].offset[arg0->count];
+    mem_ptr = *offset_ptr + (intptr_t) &arg0[0].offset[arg0->count];
     return (BKSpriteTextureBlock *)(mem_ptr);
 }
